Change parsed floats to doubles

+++ JsonParser/src/main/java/com/github/anthogis/json_parser/api/JSONParser.java
@@ -202,7 +202,7 @@ public class JSONParser {
                     addValue = true;
                     break;
                 case FLOAT:
-                    attribute = new JSONAttribute<>(key, Float.parseFloat(value));
+                    attribute = new JSONAttribute<>(key, Double.parseDouble(value));
                     addValue = true;
                     break;
                 case STRING:
